ci: measure coverage across packages and prune dead test helpers
