--- internal/identity/http/fixtures_test.go
@@ -1,17 +1,0 @@
-package identityhttp
-
-import (
-	"testing"
-
-	_ "github.com/mattn/go-sqlite3"
-)
-
-// isolateCrontab empties PATH so the `crontab` command cannot be found. The
-// backup handlers call CronManager.SyncCrontab unconditionally, which shells
-// out to crontab and would otherwise rewrite the crontab of whoever runs the
-// suite. With crontab unreachable the handlers log a warning and carry on,
-// which is the behaviour under test.
-func isolateCrontab(t *testing.T) {
-	t.Helper()
-	t.Setenv("PATH", t.TempDir())
-}
